finnhub-pipeline/requirements.txt
# Finnhub Trading Pipeline - Root Dependencies
# This file lists common dependencies used across services

# Kafka
kafka-python==2.0.2

# WebSocket
websocket-client==1.6.4

# Monitoring
prometheus-client==0.19.0

# Data Processing
pyspark==3.5.0
cassandra-driver==3.28.0

# Data Validation & Parsing
python-dotenv==1.0.0
pydantic==2.5.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.0

# Schema Registry (Optional)
confluent-kafka[avro]==2.3.0

# Data Analysis (Optional)
pandas==2.1.3
numpy==1.26.2

# Logging
python-json-logger==2.0.7

# Configuration
pyyaml==6.0.1

# Testing
requests==2.31.0
responses==0.24.1

# Note: Service-specific requirements are in:
# - finnhub_producer/requirements.txt
# - spark_processor/requirements.txt
# - requirements-dev.txt (development only)